feat：align插件增加组件props，修复移动位置时tip不跟随问题
